[feat] 우산 어드민 전체 통계, 지점 통계 구현 및 테스트 코드 작성 (#140, #141, #142) (#144)
* feat: Umbrella 도메인 전체 통계 기능 구현

* feat: Umbrella 지점 통계 기능 및 Test 코드 작성, api 업데이트

* feat: Umbrella 분실여부 필드 추가 및 테스트 코드 수정, 수정 시에 분실여부 변경 가능하도록 수정

* feat: 전체, 지점별 전체 대여건수 조회 기능 추가 구현 및 테스트 수정

* refactor: count QueryDSL로 변경

* feat: 협업지점 고유번호 예외핸들링 추가

* feat: 협업지점 고유번호 없을 시 컨트롤러 테스트 추가

* refactor: 파라미터 순서 변경

* refactor: 컨벤션 맞게 수정

+++ src/main/java/upbrella/be/store/dto/response/StoreFindByIdResponse.java
@@ -14,12 +14,12 @@ public class StoreFindByIdResponse {
     private String businessHours;
     private String contactNumber;
     private String address;
-    private int availableUmbrellaCount;
+    private long availableUmbrellaCount;
     private boolean openStatus;
     private double latitude;
     private double longitude;
 
-    public static StoreFindByIdResponse fromStoreDetail(StoreDetail storeDetail, int availableUmbrellaCount) {
+    public static StoreFindByIdResponse fromStoreDetail(StoreDetail storeDetail, long availableUmbrellaCount) {
 
         return StoreFindByIdResponse.builder()
                 .id(storeDetail.getId())
